Update Neclo_Rheslar.pl
Per alla, Note to Neclo upon server start and before February 2000 patch should yield an Elemental Armor scroll that sells for 3.5pp.  Then they changed him so that he only gives out a scroll of minor shielding.  I'm suggesting that we keep it with Elemental Armor on the Classic server, and adjust it in Kunark.